cleanup the screen viewport handling with new helpers
