+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLscreen.cpp
@@ -21,10 +21,7 @@
 #include "Graphics_Systems/OpenGL/GLscreen.h"
 #include "OpenGLHeaders.h"
 #include "Graphics_Systems/General/GSscreen.h"
-#include "Graphics_Systems/General/GSprimitives.h"
 
-#include "Universal_System/var4.h"
-#include "Universal_System/roomsystem.h"
 #include "Platforms/General/PFwindow.h"
 
 using namespace enigma;
@@ -32,14 +29,13 @@ using namespace enigma;
 namespace enigma {
 
 unsigned int bound_framebuffer = 0; //Shows the bound framebuffer, so glGetIntegerv(GL_FRAMEBUFFER_BINDING_EXT, &fbo); don't need to be called (they are very slow)
-int viewport_x, viewport_y, viewport_w, viewport_h; //These are used by surfaces, to set back the viewport
 
 void scene_begin() {
 
 }
 
 void scene_end() {
-	msaa_fbo_blit();
+  msaa_fbo_blit();
 }
 
 unsigned char* graphics_copy_screen_pixels(int x, int y, int width, int height, bool* flipped) {
@@ -70,27 +66,3 @@ unsigned char* graphics_copy_screen_pixels(unsigned* fullwidth, unsigned* fullhe
 }
 
 } // namespace enigma
-
-namespace enigma_user {
-
-void screen_set_viewport(gs_scalar x, gs_scalar y, gs_scalar width, gs_scalar height) {
-  draw_batch_flush(batch_flush_deferred);
-
-  x = (x / window_get_region_width()) * window_get_region_width_scaled();
-  y = (y / window_get_region_height()) * window_get_region_height_scaled();
-  width = (width / window_get_region_width()) * window_get_region_width_scaled();
-  height = (height / window_get_region_height()) * window_get_region_height_scaled();
-  gs_scalar sx, sy;
-  sx = (window_get_width() - window_get_region_width_scaled()) / 2;
-  sy = (window_get_height() - window_get_region_height_scaled()) / 2;
-  viewport_x = sx + x;
-  viewport_y = window_get_height() - (sy + y) - height;
-  viewport_w = width;
-  viewport_h = height;
-
-  //NOTE: OpenGL viewports are bottom left unlike Direct3D viewports which are top left
-  glViewport(viewport_x, viewport_y, viewport_w, viewport_h);
-  glScissor(viewport_x, viewport_y, viewport_w, viewport_h);
-}
-
-} // namespace enigma_user
