+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLsurface.cpp
@@ -32,12 +32,6 @@
 
 using namespace std;
 
-namespace enigma {
-
-extern int viewport_x, viewport_y, viewport_w, viewport_h;
-
-} // namespace enigma
-
 namespace enigma_user {
 
 bool surface_is_supported()
@@ -141,12 +135,10 @@ void surface_set_target(int id)
   //This fixes several consecutive surface_set_target() calls without surface_reset_target.
   int prevFbo;
   glGetIntegerv(GL_FRAMEBUFFER_BINDING_EXT, &prevFbo);
-  if (prevFbo != 0) { glPopAttrib(); glPopMatrix(); d3d_projection_stack_pop(); }
+  if (prevFbo != 0) { d3d_projection_stack_pop(); }
   get_surface(surf,id);
   glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, surf.fbo); //bind it
-  glPushMatrix(); //So you can pop it in the reset
   d3d_projection_stack_push();
-  glPushAttrib(GL_VIEWPORT_BIT); //same
   glViewport(0, 0, surf.width, surf.height);
   glScissor(0, 0, surf.width, surf.height);
   d3d_set_projection_ortho(0, surf.height, surf.width, -surf.height, 0);
@@ -157,11 +149,8 @@ void surface_reset_target(void)
   draw_batch_flush(batch_flush_deferred);
 
   glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, 0);
-  glPopAttrib();
-  glPopMatrix();
   d3d_projection_stack_pop();
-  glViewport(enigma::viewport_x, enigma::viewport_y, enigma::viewport_w, enigma::viewport_h);
-  glScissor(enigma::viewport_x, enigma::viewport_y, enigma::viewport_w, enigma::viewport_h);
+  screen_reset_viewport();
 }
 
 int surface_get_target()
